Split TaskListViewModel into multiple ViewModels
