+++ app/src/main/java/com/example/tasks/ui/viewmodel/TaskListViewModel.kt
@@ -1,21 +1,18 @@
-package com.example.tasks.ui.theme.viewmodel
+package com.example.tasks.ui.viewmodel
 
-import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.tasks.data.Category
 import com.example.tasks.data.Task
 import com.example.tasks.data.TaskDao
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import kotlinx.datetime.Clock
 import javax.inject.Inject
 
 @HiltViewModel
-class TaskViewModel @Inject constructor(private val dao: TaskDao) : ViewModel() {
+class TaskListViewModel @Inject constructor(private val dao: TaskDao) : ViewModel() {
     private var sortedByFutureDeadline =
         dao.getAllFutureByDeadlineAsc(Clock.System.now().toEpochMilliseconds())
     private var sortedByOverdueDeadline =
@@ -76,7 +73,6 @@ class TaskViewModel @Inject constructor(private val dao: TaskDao) : ViewModel() 
         }
     }
 
-    //reminder: do NOT mutate lists passed to DiffUtil, create a new list instead
     private var deadline = mutableListOf(
         Category("Overdue", overdue),
         Category("Future", future),
@@ -116,40 +112,9 @@ class TaskViewModel @Inject constructor(private val dao: TaskDao) : ViewModel() 
         }
     }
 
-    fun addTask(task: Task) {
-        Log.d("TaskViewModel", "Added task with timestamp ${task.deadline}")
-        if (task.taskName.isBlank()) task.taskName = "Untitled"
-        viewModelScope.launch {
-            dao.insert(task)
-        }
-    }
-
-    fun deleteTask(task: Task) {
-        Log.d("TaskViewModel", "deleted task with id ${task.taskId}")
-        viewModelScope.launch {
-            dao.delete(task)
-        }
-    }
-
-    fun editTask(task: Task) {
-        Log.d("TaskViewModel", "edited task with timestamp ${task.deadline}")
-        if (task.taskName.isBlank()) task.taskName = "Untitled"
-        viewModelScope.launch {
-            dao.update(task)
-        }
-    }
-
     fun completeTask(taskId: Long) {
         viewModelScope.launch {
             dao.toggleTaskDone(taskId)
         }
-    }
-
-    fun retrieveTask(taskId: Long): Flow<Task> {
-        return dao.get(taskId)
-    }
-
-    fun retrieveNextTask(present: Long): Flow<Task> {
-        return dao.getNextDeadline(present)
     }
 }
